Re-organize the code.
