+++ FbxCppTests/TestRunner.cpp
@@ -1,6 +1,7 @@
 ﻿
 #include "Tests.h"
 #include <vector>
+#include <iostream>
 
 using namespace std;
