+++ FbxCppTests/properties.h
@@ -1,0 +1,10 @@
+
+
+#ifndef __FBXCPPTESTS_PROPERTIES_H
+#define __FBXCPPTESTS_PROPERTIES_H
+
+#include <fbxsdk.h>
+
+const char* GetTypeName(EFbxType tid);
+
+#endif // __FBXCPPTESTS_PROPERTIES_H
