+++ FbxCppTests/textures.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 #include <fbxsdk.h>
-#include "common.h"
+#include "print.h"
 
 using namespace std;
